OPlusExtras: Drop getFileValueAsBoolean usage
'getFileValueAsBoolean' assumes that the value returned by the node is always
either 0 (false) or 1 and above (true), which may not always be the case.
This can result in instances where toggles are falsely switched on.
Additionally, some nodes do not return the updated value when a 1 is
written to them, such as powershare on op8 series. To prevent these
issues, direct retrieval from sharedPrefs is used instead.

Signed-off-by: AnierinB <anierin@evolution-x.org>